% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLTM.R
\name{MLTM}
\alias{MLTM}
\title{The Multidimensional Latent Trait Model}
\usage{
GMLTM(data, Q, components, iters = 2000, iter_warmup = 1000, chains = 2, quantiles = c(0.025, 0.50,  0.975))
}
\arguments{
\item{data}{\eqn{n \times p} data.frame or data matrix with the individuals in rows and items in columns.}

\item{Q}{\eqn{p \times q} Q matrix.}

\item{components}{List of \eqn{d} elements relating each components to a vector of rules.}

\item{iters}{Number of samples from the posterior distribution.}

\item{chains}{Number of Markov chains.}

\item{iter_warmup}{Number of samples to discard then initializing a Markov chain.}

\item{quantiles}{Vector of probabilities associated with the posterior quantiles.}

\item{parallel_chains}{Number of chains to run in parallel cores.}

\item{threads_per_chain}{Number of cores to run within a chain.}

\item{...}{Additional arguments to pass to the \code{sample} function from \code{cmdstanr}.}
}
\value{
\code{cmdstan} object:
}
\description{
Estimate the parameters of the MLTM-D
}
\details{
\code{MLTM} Estimates the Bayesian version of the Multicomponent Latent Trait Model for Diagnosis (MLTM-D) by Embretson & Yang (2013), a noncompensatory latent trait model. MLTM-D specifies a hierarchical relationship between components and attributes, allowing for diagnosis at two levels. This model is applicable to broad trait measures, such as achievement tests, where the component structure varies between items and the items are composed of different cognitive operations.
}
\examples{
\dontrun{
# Load the data
data(analogy)
# These data correspond to the study:
# Blum, D., Holling, H., Galibert, M. S., & Forthmann, B. (2016). Task difficulty prediction of figural analogies. Intelligence, 56, 72-81.
# Define the dataset

dataset <- analogy

# Define the Q matrix with item rules
Q <- structure(c(0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1,
                1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1,
                0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0,
                0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1,
                0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1,
                0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1,
                1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1), dim = c(27L, 5L), dimnames = list(
                  c("item 1", "item 2", "item 3", "item 4", "item 5", "item 6",
                    "item 7", "item 8", "item 9", "item 10", "item 11", "item 12",
                    "item 13", "item 14", "item 15", "item 16", "item 17", "item 18",
                    "item 19", "item 20", "item 21", "item 22", "item 23", "item 24",
                    "item 25", "item 26", "item 27"), c("rot_fig", "rot_trap",
                                                        "reflection", "subt_seg", "mov_point"))) #The labels correspond to the names of the rules.


# Define the components list
components <- list(global = c(1, 2, 3), local = c(4, 5))

# Define the fit2 object
fit2 <- MLTM(dataset, Q, components, iters = 2000, iter_warmup = 500,
             quantiles = c(0.05, 0.5, 0.95), chains = 2, parallel_chains = 2)

# Print the difficulty parameters for each rule in each component
fit2$EAP$eta

# Print the difficulty parameter for each item. For items with rules from both components, print both difficulties.
fit2$EAP$beta

# Print the discriminations of each item. When an item is composed of rules from both components, provide a discrimination for each component.
fit2$EAP$alpha

# Provides a summary of parameter estimates and diagnoses convergence in a Bayesian model for the eta parameter.
fit2$fit$print("eta")

# Provides a summary of parameter estimates and diagnoses convergence in a Bayesian model for the alpha parameter.
fit2$fit$print("alpha")

# Marginal reliability for each component
reliability(fit2)

# Print all ppchecks of the model
check2 <- ppchecks(fit2)
}

}
\references{
Embretson, S. E., & Yang, X. (2013). A multicomponent latent trait model for diagnosis. \emph{Psychometrika}, 78, 14-36.\cr

Ramírez, E.S.; Jiménez, M.; Franco, V.R.; Alvarado, J.M. Delving into the Complexity of Analogical Reasoning: A Detailed Exploration with the Generalized Multicomponent Latent Trait Model for Diagnosis. \emph{J. Intell.} 2024, 12, 67. https://doi.org/10.3390/jintelligence12070067
}
