% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppchecks.R
\name{ppchecks}
\alias{ppchecks}
\title{Posterior predictive checks, fitted values, and prediction intervals.}
\usage{
ppchecks(fit, nsim = 100, interval = 0.95)
}
\arguments{
\item{fit}{MLTM object.}

\item{nsim}{Number of simulated samples.}

\item{interval}{Probability associated with the credibility intervals.}

\item{...}{Additional arguments to pass to the \code{plot} function.}
}
\value{
An array of simulated scores.
}
\description{
Plot the frequency distribution of the empirical and simulated total scores and obtain the fitted values and prediction intervals.
}
\details{
\code{ppchecks} estimates a ...
}
\references{
Ramírez, E.S.; Jiménez, M.; Franco, V.R.; Alvarado, J.M. Delving into the Complexity of Analogical Reasoning: A Detailed Exploration with the Generalized Multicomponent Latent Trait Model for Diagnosis. \emph{J. Intell.} 2024, 12, 67. https://doi.org/10.3390/jintelligence12070067
}
