% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability}
\alias{reliability}
\title{Marginal reliability}
\usage{
reliability(fit)
}
\arguments{
\item{fit}{MLTM object.}
}
\value{
A number denoting the reliability estimate.
}
\description{
Estimate the the marginal reliability of the GMLTM.
}
\details{
\code{reliability} estimates a ...
}
\references{
Ramírez, E.S.; Jiménez, M.; Franco, V.R.; Alvarado, J.M. Delving into the Complexity of Analogical Reasoning: A Detailed Exploration with the Generalized Multicomponent Latent Trait Model for Diagnosis. \emph{J. Intell.} 2024, 12, 67. https://doi.org/10.3390/jintelligence12070067
}
